requirements.txt
# Core transcription
faster-whisper>=1.0.0

# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# System tray
pystray>=0.19.0
Pillow>=10.0.0

# Global hotkeys
keyboard>=0.13.5

# Text output
pyautogui>=0.9.54
pyperclip>=1.8.0

# Notifications
plyer>=2.1.0

# Testing
pytest>=7.0.0
